Move shields to api folder in helpers
