+++ src/lib/components/mod-info/mod-info.svelte
@@ -82,7 +82,7 @@
 				</a>
 			</div>
 			<div>
-				<a class="link" href={`/api/${mod.uniqueName}/badge.json`}> Get Install Badge </a>
+				<a class="link" href="/api/{mod.uniqueName}/badge.json"> Get Install Badge </a>
 			</div>
 			<div>
 				<code
