+++ src/routes/api/[modUniqueName]/badge.json/+server.ts
@@ -1,6 +1,6 @@
 import { error, type RequestHandler } from '@sveltejs/kit';
 import type { Mod } from '$lib/helpers/api/get-mod-list';
-import { BASE_BADGE, type ShieldsEndpointBadge } from '../../../../lib/helpers/shields';
+import { BASE_BADGE, type ShieldsEndpointBadge } from '$lib/helpers/api/shields';
 
 type Params = {
     modUniqueName: string;
